// entidad-web/entidad-web/src/app/@data/model/reporte.ts
export interface IConvocatoria {
  correlativo: number;
  entidadId: number;
  nombreEntidad: string;
  baseId: number;
  estadoBaseId: number;
  estadoBase: string;
  fechaCreacionBase: string;
  convocatoriaId: number;
  gestorId: number;
  nombreGestor: string;
  coordinadorId: number;
  nombreCoordinador: string;
  estadoConvocatoriaId: number;
  estadoConvocatoria: string;
  etapaConvocatoriaId: number;
  etapaConvocatoria: string;
  regimenId: number;
  regimen: string;
  modalidadId: number;
  modalidad: string;
  tipoId: number;
  tipo: string;
  denominacionPuesto: string;
  nroVacantes: number;
  tipoPractica: string;
  nroPostPorConvo: number;
  nroPostContratados: number;
  fechaPublicacion: string;
  codigoConvocatoria: string;
  periodoIniEtapa: string;
  periodoFinEtapa: string;
  departamentoId: number;
  descDepartamento: string;
  reportePerfilId: number;
  descripcionPerfil: string;
  remuneracion: number;
}